Split more modules from lib/url.
Extracted `is-outside-calypso`, `is-https` and `scheme-utils`.
